[🔧수정] AI28Layer 컴포넌트 리팩토링
