+++ src/components/index.ts
@@ -11,7 +11,7 @@ export { default as Input } from './Input/Input'
 export { default as Result } from './Result/Result'
 export { default as Navigation } from './Navigation/Navigation'
 export { default as Pagination } from './Pagination/Pagination'
-export { default as Layer } from './Layer/Layer'
+export { default as AI28Layer } from './AI28Layer/AI28Layer'
 export { default as Layers } from './Layers/Layers'
 export { default as Table } from './Table/Table'
 export { default as Canvas } from './Canvas/Canvas'
